+++ packages/main/src/models/generative-model.test.ts
@@ -18,7 +18,9 @@ import { expect, use } from "chai";
 import { GenerativeModel } from "./generative-model";
 import * as sinonChai from "sinon-chai";
 import {
+  CountTokensRequest,
   FunctionCallingMode,
+  FunctionDeclarationSchemaType,
   HarmBlockThreshold,
   HarmCategory,
 } from "../../types";
@@ -50,7 +52,19 @@ describe("GenerativeModel", () => {
       "apiKey",
       {
         model: "my-model",
-        generationConfig: { temperature: 0 },
+        generationConfig: {
+          temperature: 0,
+          responseMimeType: "application/json",
+          responseSchema: {
+            type: FunctionDeclarationSchemaType.OBJECT,
+            properties: {
+              testField: {
+                type: FunctionDeclarationSchemaType.STRING,
+                properties: {},
+              },
+            },
+          },
+        },
         safetySettings: [
           {
             category: HarmCategory.HARM_CATEGORY_DANGEROUS_CONTENT,
@@ -68,6 +82,15 @@ describe("GenerativeModel", () => {
       },
     );
     expect(genModel.generationConfig?.temperature).to.equal(0);
+    expect(genModel.generationConfig?.responseMimeType).to.equal(
+      "application/json",
+    );
+    expect(genModel.generationConfig?.responseSchema.type).to.equal(
+      FunctionDeclarationSchemaType.OBJECT,
+    );
+    expect(
+      genModel.generationConfig?.responseSchema.properties.testField.type,
+    ).to.equal(FunctionDeclarationSchemaType.STRING);
     expect(genModel.safetySettings?.length).to.equal(1);
     expect(genModel.tools?.length).to.equal(1);
     expect(genModel.toolConfig?.functionCallingConfig.mode).to.equal(
@@ -92,6 +115,7 @@ describe("GenerativeModel", () => {
           value.includes(FunctionCallingMode.NONE) &&
           value.includes("be friendly") &&
           value.includes("temperature") &&
+          value.includes("testField") &&
           value.includes(HarmBlockThreshold.BLOCK_LOW_AND_ABOVE)
         );
       }),
@@ -129,7 +153,19 @@ describe("GenerativeModel", () => {
   it("generateContent overrides model values", async () => {
     const genModel = new GenerativeModel("apiKey", {
       model: "my-model",
-      generationConfig: { temperature: 0 },
+      generationConfig: {
+        temperature: 0,
+        responseMimeType: "application/json",
+        responseSchema: {
+          type: FunctionDeclarationSchemaType.OBJECT,
+          properties: {
+            testField: {
+              type: FunctionDeclarationSchemaType.STRING,
+              properties: {},
+            },
+          },
+        },
+      },
       safetySettings: [
         {
           category: HarmCategory.HARM_CATEGORY_DANGEROUS_CONTENT,
@@ -152,7 +188,18 @@ describe("GenerativeModel", () => {
       mockResponse as Response,
     );
     await genModel.generateContent({
-      generationConfig: { topK: 1 },
+      generationConfig: {
+        topK: 1,
+        responseSchema: {
+          type: FunctionDeclarationSchemaType.OBJECT,
+          properties: {
+            newTestField: {
+              type: FunctionDeclarationSchemaType.STRING,
+              properties: {},
+            },
+          },
+        },
+      },
       safetySettings: [
         {
           category: HarmCategory.HARM_CATEGORY_HARASSMENT,
@@ -179,6 +226,8 @@ describe("GenerativeModel", () => {
           value.includes(FunctionCallingMode.AUTO) &&
           value.includes("be formal") &&
           value.includes("topK") &&
+          value.includes("newTestField") &&
+          !value.includes("testField") &&
           value.includes(HarmCategory.HARM_CATEGORY_HARASSMENT)
         );
       }),
@@ -203,7 +252,6 @@ describe("GenerativeModel", () => {
       mockResponse as Response,
     );
     await genModel.countTokens("hello");
-    console.log(makeRequestStub.args[0]);
     expect(makeRequestStub).to.be.calledWith(
       "models/my-model",
       request.Task.COUNT_TOKENS,
@@ -254,9 +302,24 @@ describe("GenerativeModel", () => {
   it("passes params through to chat.sendMessage", async () => {
     const genModel = new GenerativeModel("apiKey", {
       model: "my-model",
+      generationConfig: {
+        temperature: 0,
+        responseMimeType: "application/json",
+        responseSchema: {
+          type: FunctionDeclarationSchemaType.OBJECT,
+          properties: {
+            testField: {
+              type: FunctionDeclarationSchemaType.STRING,
+              properties: {},
+            },
+          },
+        },
+      },
       systemInstruction: { role: "system", parts: [{ text: "be friendly" }] },
     });
     expect(genModel.systemInstruction?.parts[0].text).to.equal("be friendly");
+    expect(genModel.generationConfig.responseSchema.properties.testField).to
+      .exist;
     const mockResponse = getMockResponse(
       "unary-success-basic-reply-short.json",
     );
@@ -270,7 +333,7 @@ describe("GenerativeModel", () => {
       match.any,
       false,
       match((value: string) => {
-        return value.includes("be friendly");
+        return value.includes("be friendly") && value.includes("testField");
       }),
       {},
     );
@@ -279,10 +342,25 @@ describe("GenerativeModel", () => {
   it("startChat overrides model values", async () => {
     const genModel = new GenerativeModel("apiKey", {
       model: "my-model",
+      generationConfig: {
+        temperature: 0,
+        responseMimeType: "application/json",
+        responseSchema: {
+          type: FunctionDeclarationSchemaType.OBJECT,
+          properties: {
+            testField: {
+              type: FunctionDeclarationSchemaType.STRING,
+              properties: {},
+            },
+          },
+        },
+      },
       tools: [{ functionDeclarations: [{ name: "myfunc" }] }],
       toolConfig: { functionCallingConfig: { mode: FunctionCallingMode.NONE } },
       systemInstruction: { role: "system", parts: [{ text: "be friendly" }] },
     });
+    expect(genModel.generationConfig.responseSchema.properties.testField).to
+      .exist;
     expect(genModel.tools?.length).to.equal(1);
     expect(genModel.toolConfig?.functionCallingConfig.mode).to.equal(
       FunctionCallingMode.NONE,
@@ -297,6 +375,17 @@ describe("GenerativeModel", () => {
     await genModel
       .startChat({
         tools: [{ functionDeclarations: [{ name: "otherfunc" }] }],
+        generationConfig: {
+          responseSchema: {
+            type: FunctionDeclarationSchemaType.OBJECT,
+            properties: {
+              newTestField: {
+                type: FunctionDeclarationSchemaType.STRING,
+                properties: {},
+              },
+            },
+          },
+        },
         toolConfig: {
           functionCallingConfig: { mode: FunctionCallingMode.AUTO },
         },
@@ -312,11 +401,43 @@ describe("GenerativeModel", () => {
         return (
           value.includes("otherfunc") &&
           value.includes(FunctionCallingMode.AUTO) &&
-          value.includes("be formal")
+          value.includes("be formal") &&
+          value.includes("newTestField") &&
+          !value.includes("testField")
         );
       }),
       {},
     );
+    restore();
+  });
+  it("countTokens errors if contents and generateContentRequest are both defined", async () => {
+    const genModel = new GenerativeModel(
+      "apiKey",
+      {
+        model: "my-model",
+      },
+      {
+        apiVersion: "v2000",
+      },
+    );
+    const mockResponse = getMockResponse(
+      "unary-success-basic-reply-short.json",
+    );
+    const makeRequestStub = stub(request, "makeRequest").resolves(
+      mockResponse as Response,
+    );
+    const countTokensRequest: CountTokensRequest = {
+      contents: [{ role: "user", parts: [{ text: "hello" }] }],
+      generateContentRequest: {
+        contents: [{ role: "user", parts: [{ text: "hello" }] }],
+      },
+    };
+    await expect(
+      genModel.countTokens(countTokensRequest),
+    ).to.eventually.be.rejectedWith(
+      "CountTokensRequest must have one of contents or generateContentRequest, not both.",
+    );
+    expect(makeRequestStub).to.not.be.called;
     restore();
   });
 });