+++ packages/main/src/requests/request-helpers.ts
@@ -17,11 +17,16 @@
 
 import {
   Content,
+  CountTokensRequest,
+  CountTokensRequestInternal,
   EmbedContentRequest,
   GenerateContentRequest,
   Part,
 } from "../../types";
-import { GoogleGenerativeAIError } from "../errors";
+import {
+  GoogleGenerativeAIError,
+  GoogleGenerativeAIRequestInputError,
+} from "../errors";
 
 export function formatSystemInstruction(
   input?: string | Part | Content,
@@ -102,6 +107,31 @@ function assignRoleToPartsAndValidateSendMessageRequest(
   }
 
   return functionContent;
+}
+
+export function formatCountTokensInput(
+  params: CountTokensRequest | string | Array<string | Part>,
+  model: string,
+): CountTokensRequestInternal {
+  let formattedRequest: CountTokensRequestInternal = {};
+  const containsGenerateContentRequest =
+    (params as CountTokensRequest).generateContentRequest != null;
+  if ((params as CountTokensRequest).contents) {
+    if (containsGenerateContentRequest) {
+      throw new GoogleGenerativeAIRequestInputError(
+        "CountTokensRequest must have one of contents or generateContentRequest, not both.",
+      );
+    }
+    formattedRequest = { ...(params as CountTokensRequest) };
+  } else if (containsGenerateContentRequest) {
+    formattedRequest = { ...(params as CountTokensRequest) };
+    formattedRequest.generateContentRequest.model = model;
+  } else {
+    // Array or string
+    const content = formatNewContent(params as string | Array<string | Part>);
+    formattedRequest.contents = [content];
+  }
+  return formattedRequest;
 }
 
 export function formatGenerateContentInput(